+++ config/index.js
@@ -18,7 +18,7 @@ const defaultRoot = resolve(process.env.HOME, ".cynomys", "protal");
 
 const [num, dataDir, logDir] = [cpus().length, resolve(defaultRoot, "data"), resolve(defaultRoot, "logs")];
 
-siteConfig.set("version", "1.0.8"); // 设置服务器版本
+siteConfig.set("version", "v1.0.9bate"); // 设置服务器版本
 siteConfig.set("ssl", { // 安全设置
   enable: true, // 是否开启ssl设置
   keyName: "cynomy", // ssl公私钥 文件名 // 不写则默认